perf: fuse dQ reduction and fp16 output

+++ include/fused_bwd_dq_reduce.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <hip/hip_fp16.h>
+
 #include "shaobo_fa3_api.h"
 
 #include <cstddef>
@@ -9,7 +11,7 @@ namespace shaobo::fa3::bwd::fused_bwd {
 size_t dq_workspace_bytes(const ShaoboFa3Params* params);
 
 int launch_dq_reduction(const float* partial,
-                        float* dq,
+                        __half* dq,
                         const ShaoboFa3Params* params);
 
 }  // namespace shaobo::fa3::bwd::fused_bwd
